requirements.txt
# Python dependencies for the caffeine estimation pipeline
# Install with:  pip install -r requirements.txt

pyserial==3.5
numpy>=1.24
scipy>=1.13
pandas>=2.0
neurokit2==0.2.9
matplotlib>=3.7
